+++ RELEASE_NOTES.md
@@ -1,5 +1,12 @@
 # Pushrod Releases
 
+## 0.4.6
+
+- Added on_mouse_button callback (#184, #185)
+- Added on_tick callback (#187)
+- Added on_mouse_bounds callback (#181, #182)
+- Added inject_event_handler macro to handle default events for callbacks (#192)
+
 ## 0.4.5
 
 - Changed get_widget_position_by_name to get_widget_by_name (far easier to use)